Complete overhaul. Now using C++20.
The parser is now constructed using static std::string_view's and std::array's instead of dynamic std::string's and std::vector's.
Improved performance: now an option does all the work to determine whether to assign itself to the provided argument, and returns true when it does. So argument are not searched twice, as was done before (see old OptionBase::operator= and operator==).
Now the value is saved inside a provided output reference, so that there is no need to find it using a string as a key.
Now there are 3 types of Option:
 - SwitchOption: requires no value and sets its output to a specified value.
 - ManualOption: requires a value and sets its output based on a user-defined string-to-output-type conversion function.
 - Option: requires a value and sets its output by automatically converting the provided string to the type of the output (e.g. when the output is an int, the provided string is interpreded as a number).
Instead of using a ArgParser::parse(std::vector<std::string>) overload for arguments not coming from main(argc, argv), added an abstract parse(first, last) with iterators.
Now the user has to specify if firstArgumentIsExecutablePath in ArgParser::parse() and not in the constructor (defaults to true for the parse(int, char const* []) overload).
In help screen, now adding a newline after the arguments when the description indentation would be bigger than the requested indentation.
Added HelpSection class to be able to add titles to the help screen for various sections directly from the ArgParser constructor.